Updated our project structure to align it properly with navigating and routing.
