--- Red River Mart/src/components/common/headerPage.tsx
@@ -1,20 +1,0 @@
-import "./header.css";
-import rrmLogo from "../../assets/rrmLogo.png";
-export default function Header() {
-  return (
-    <header className="header">
-      <div className="logo">
-        <img src={rrmLogo} alt="Red River Mart Logo" />
-      </div>
-
-      <nav className="menu">
-        <ul>
-          <li>Buy</li>
-          <li>Sell</li>
-          <li>Inbox</li>
-          <li>Account</li>
-        </ul>
-      </nav>
-    </header>
-  );
-}
