+++ Red River Mart/src/components/layout/layout.tsx
@@ -1,0 +1,14 @@
+import { Outlet } from "react-router-dom";
+import Footer from "./footer/footerPage";
+import Nav from "./nav/navBar";
+
+export function Layout() { 
+    return(
+        <>
+        <Nav />
+        <Outlet />
+        <Footer />
+        
+        </>
+    )
+}
