+++ Red River Mart/src/components/pages/sellComponent/sellPage.css
